encoding/wkb,encoding/ewkb: fix encoding for POINT EMPTY
For POINT EMPTY, NaN is used as the XYZM values in WKB/EWKB encoding.

See Requirement 152 of http://www.geopackage.org/spec/:

```
If the geometry is a Point, it SHALL be encoded with each coordinate value set to an IEEE-754 quiet NaN value. GeoPackages SHALL use big endian 0x7ff8000000000000 or little endian 0x000000000000f87f as the binary encoding of the NaN values. (This is because Well-Known Binary as defined in OGC 06-103r4 [9] does not provide a standardized encoding for an empty point set, i.e., 'Point Empty' in Well-Known Text.)
```

This commit addresses that behaviour.