modify POINT to use [NaN] as the coordinates for empty
To follow Requirement 152 of http://www.geopackage.org/spec/:

```
If the geometry is a Point, it SHALL be encoded with each coordinate value set to an IEEE-754 quiet NaN value. GeoPackages SHALL use big endian 0x7ff8000000000000 or little endian 0x000000000000f87f as the binary encoding of the NaN values. (This is because Well-Known Binary as defined in OGC 06-103r4 [9] does not provide a standardized encoding for an empty point set, i.e., 'Point Empty' in Well-Known Text.)
```

and allow `MULTIPOINT (EMPTY, POINT(1.0 1.0))` to be represented, we
need to convert POINT to use NaN to encode empty.

Ran into reflect.DeepEqual not respecting NaN == NaN, so had to split
some tests out of their shell.