fix MultiPolygon and MultiLineString to handle EMPTY sub-geometries
MultiLineString / MultiPolygon should be to handle extreme cases such as
MULTILINESTRING (EMPTY, EMPTY) or MULTIPOLYGON (EMPTY, POLYGON(1.0...)).
This PR adapts the code to make this happen.

Unfortunately, it will not work for MultiPoint. I'm guessing this is the
reason the spec has:

```
If the geometry is a Point, it SHALL be encoded with each coordinate value set to an IEEE-754 quiet NaN value. GeoPackages SHALL use big endian 0x7ff8000000000000 or little endian 0x000000000000f87f as the binary encoding of the NaN values. (This is because Well-Known Binary as defined in OGC 06-103r4 [9] does not provide a standardized encoding for an empty point set, i.e., 'Point Empty' in Well-Known Text.)
```

It is impossible to tell between "empty" coordinates in MultiPoint if we
store it as an array of coordinates. This may render #159
incorrect -- we should instead handle empty coordinates as (NaN, NaN).