encoding/*wkb: fix panic on invalid layout on empty geometries
Previously, if you had `geom.NewPointFlat(geom.NoLayout)`, and tried to
cast it to a WKB or EWKB, a panic would occur instead of invalid layout,
This is because the `g` argument would become `nil` and hence
`g.Layout()` would be a nil dereference. Fix that up by not reassigning
`g` (as Empty() is now part of the interface so we don't need it
dereferenced).