encoding/wkt: expose maxDecimalDigits as an option
This commit adds support for maximum decimal digits in WKT encoding,
similar to the way PostGIS handles ST_AsText(<geom>, <maxdecimaldigits>).
We have to do extra processing to remove trailing 0 and . marks.

To make this change, we also expose "Encoder" as a struct, with the
Encode object made public. This can take in `EncodeOption`s which allow
an arbitrary amount of options in a clean way.