encoding/wkb: fix misleading error message
wkbcommon.ErrUnexpectedType.Got is filled with the result of the type
assertion, while it should receive the source of it.

When the type assertion failed, the first returned value is just an
empty expected type. Using it in Got makes error message like :

wkb: got *geom.Point, want *wkb.Point

which is really misleading. With this commit, the error message looks
like :

wkb: got *geom.Polygon, want *wkb.Point

which has a lot more sense, we immediately understand the problem is
because of wrong geography kinds.